added coversion to double

+++ common/playacem/allrondism/block/BlockContainerAM.java
@@ -53,7 +53,7 @@ public abstract class BlockContainerAM extends BlockContainer {
     public void onBlockPlacedBy(World world, int x, int y, int z, EntityLiving entityLiving, ItemStack itemStack) {
 
         int direction = 0;
-        int facing = MathHelper.floor_double(entityLiving.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
+        int facing = MathHelper.floor_double((double)entityLiving.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 
         if (facing == 0) {
             direction = ForgeDirection.NORTH.ordinal(); // 2
